RealPathUtilV2
This is our own custom build real path utility which doesn't rely on content resolvers and media queries to retrieve sharable content. Instead this class will create a temporary file in the users cache documents directory and then return the link to the newly created file. This seems to work on all android versions for downloads, media, Google drive, etc.

The downside to doing it this way is that extra time is spent copying the file locally and then uploading the copy, which for large video files may take twice as long. However, this method seems to be far more robust and easier to maintain.